moving files to common cacher folder
