+++ gtsfm/configs/sift_front_end.yaml
@@ -8,7 +8,7 @@ SceneOptimizer:
   feature_extractor:
     _target_: gtsfm.feature_extractor.FeatureExtractor
     detector_descriptor:
-      _target_: gtsfm.frontend.detector_descriptor.cacher.detector_descriptor_cacher.DetectorDescriptorCacher
+      _target_: gtsfm.frontend.cacher.detector_descriptor_cacher.DetectorDescriptorCacher
       detector_descriptor_obj:
         _target_: gtsfm.frontend.detector_descriptor.sift.SIFTDetectorDescriptor
 
@@ -19,7 +19,7 @@ SceneOptimizer:
     min_num_inliers_acceptance: 15
     
     matcher:
-      _target_: gtsfm.frontend.matcher.cacher.matcher_cacher.MatcherCacher
+      _target_: gtsfm.frontend.cacher.matcher_cacher.MatcherCacher
       matcher_obj:
         _target_: gtsfm.frontend.matcher.twoway_matcher.TwoWayMatcher
