+++ gtsfm/frontend/cacher/detector_descriptor_cacher.py
